backend/pyproject.toml
[project]
name = "excelflow-backend"
version = "0.1.0"
description = "Excel/LibreOffice Template Mapper MVP"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "SQLAlchemy>=2.0.30",
  "psycopg[binary]>=3.1.19",
  "alembic>=1.13.2",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "openpyxl>=3.1.5",
  "pypdf>=4.2.0",
  "reportlab>=4.2.0",
]
